made scripts folder, updated docs, deleted TimeseriesData class
