+++ docs/models.md
@@ -15,7 +15,11 @@ An enum class for defining different ways to sample our timeseries data accordin
 An enum class for defining different severity levels of rules. For example, a rule checking for CO2 levels of over 1000 ppm might have a severity of HIGH, while a rule checking for over 1500 ppm might have a severity level of CRITICAL.
 
 ### Condition
-Represents the specific details of a rule. Contains a `to_dict()` method to easily make it into a json object. Threshold can either be an integer or a tuple containing two integers that represent a range. 
+Represents the specific details of a rule. Contains a `to_dict()` method to easily make it into a json object. 
+* Threshold: can either be an integer or a tuple containing two integers that represent a range. 
+* sleep_time: how often the rule is run
+* duration: the window size of the rolling average, the length of time that is considered an anomaly
+the length of time from which the average is calculated from 
 
 ### Anomaly
 Represents an anomaly. Contains a `to_tuple()` method to easily format and insert into Postgres table:
@@ -33,6 +37,3 @@ start_time              |        end_time        | rule_id | anomaly_id |  value
 
 ### PointReading
 Defines the format of individual timeseries data being inserted into the Postgres `timeseries` table. Only used for creating sample data.
-
-### TimeseriesData
-Not used at the moment.
